Initializer in DB

+++ PriceCompare/PriceCompare/Data/Chain.cs
@@ -4,7 +4,7 @@ namespace Data
 {
     public class Chain
     {
-        public int ChainID { get; set; }
+        public long ChainID { get; set; }
         public string Name { get; set; }
 
         public virtual ICollection<Store> Stores { get; set; }
